+++ src/test/java/io/github/astrapi69/crypt/api/algorithm/MdAlgorithmTest.java
@@ -28,9 +28,6 @@ import static org.junit.jupiter.api.Assertions.assertEquals;
 
 import org.junit.jupiter.api.Test;
 
-import io.github.astrapi69.crypt.api.algorithm.compound.CompoundAlgorithm;
-import io.github.astrapi69.crypt.api.algorithm.compound.CompoundAlgorithmTest;
-
 /**
  * The test class {@link MdAlgorithmTest} provides unit tests for the {@link MdAlgorithm} enum.
  */