+++ src/test/java/io/github/astrapi69/crypt/api/obfuscation/rule/OperationTest.java
@@ -36,8 +36,6 @@ import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.CsvFileSource;
 import org.junit.jupiter.params.provider.MethodSource;
 
-import io.github.astrapi69.crypt.api.obfuscation.rule.Operation;
-
 /**
  * The unit test class for the class {@link Operation}
  */